modules for server-admin-page v3.1 update 1 csrf prevetion the php server can be started by a non-root user prevent-direct optimization css cache headers css and js minification opcache control library

+++ www/power/index.php
@@ -1,12 +1,14 @@
 <?php include($system['location_php'] . '/lib/login/login.php'); ?>
+<?php chdir($system['location_php']); ?>
+<?php if(!csrf_checkToken('get')){ include($system['location_php'] . '/lib/prevent-index.php'); exit(); } ?>
 <?php
 	switch($_GET['do'])
 	{
 		case 'halt':
-			shell_exec('halt');
+			shell_exec($system['location_php'] . strtok($_SERVER['REQUEST_URI'], '?') . '/shell.sh halt');
 			break;
 		case 'reboot':
-			shell_exec('reboot');
+			shell_exec($system['location_php'] . strtok($_SERVER['REQUEST_URI'], '?') . '/shell.sh reboot');
 			break;
 		case 'suspend':
 			shell_exec('nohup ' . $system['location_php'] . strtok($_SERVER['REQUEST_URI'], '?') . '/shell.sh suspend > /dev/null 2>&1 &');
@@ -18,7 +20,6 @@
 	<head>
 		<title>Shutdown</title>
 		<?php include($system['location_php'] . '/lib/htmlheaders.php'); ?>
-		<?php include 'favicon.php'; ?>
 	</head>
 	<body>
 		<h1>Doing <?php echo $_GET['do']; ?>...</h1>
